requirements.txt
boto3==1.35.26
Flask==3.0.3
Flask_Cors==5.0.0
flask_wtf==1.2.1
python-dotenv==1.0.1
Requests==2.32.3
